requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==4.1.7
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-spectacular==0.26.0
drf-spectacular-sidecar==2023.3.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
